Finished Project

+++ G00332787/src/main/java/com/ships/repositories/ShippingCompanyRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.ships.model.ShippingCompany;
 
-public interface ShippingCompanyRepository extends CrudRepository<ShippingCompany, Long> {
+public interface ShippingCompanyRepository extends CrudRepository<ShippingCompany, Integer> {
 	
 }
